Rename TaskAuthenticate -> TaskBrowserAuth
